+++ vite.config.ts
@@ -1,9 +1,8 @@
 import { defineConfig } from 'vite'
-import reactPlugin from '@vitejs/plugin-react'
 import logseqDevPlugin from 'vite-plugin-logseq'
 
 export default defineConfig({
-  plugins: [logseqDevPlugin(), reactPlugin()],
+  plugins: [logseqDevPlugin()],
   build: {
     target: 'esnext',
     minify: 'esbuild',